refactor: move to `Options`
